orchestrator/requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-cors==6.0.1
flask-socketio==5.5.1
python-socketio==5.13.0
redis==6.2.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pyyaml==6.0.2
requests==2.32.4
eventlet>=0.33.0
gunicorn>=20.1.0